Restructure of Repo
- Restructured Repository to be more organized with modules that relate to each other. Common object that will be publicly exposed are Parser, Molecule and XYZFile, this may change later
- Added a Calculation folder where the calculation classes and single Calculation types will be (Ex: GeoOpt, GOAT)
- Added a temporary Spectra File to the Pipeline folder to represent that Spectra being chained calculations is a Pipeline
- Fixed __init__.py files to properly import new changes
- Fixed Spacing
- Tested previous tests to make sure nothing broke
