+++ qchem/Calculation/GeoOpt.py
@@ -2,10 +2,10 @@ import time
 import os
 import pandas as pd
 from qchem.Molecule import Molecule
-from qchem.OrcaInputFile import OrcaInputFile
-from qchem.Data.Enums import OrcaInputTemplate, OrcaCalculationType
-from qchem.OrcaCalculation import OrcaCalculation
 from qchem.Parser import OrcaOutput
+from qchem.Calculation.OrcaInputFile import OrcaInputFile
+from qchem.Calculation.OrcaCalculation import OrcaCalculation
+from qchem.Data.Enums import OrcaInputTemplate, OrcaCalculationType
 
 class GeoOpt:
