+++ qchem/Calculation/OrcaInputFile.py
@@ -1,25 +1,29 @@
-from .Data.Enums import OrcaInputTemplate
+from ..Data.Enums import OrcaInputTemplate
 
 class OrcaInputFile:
-    # Template is an input file with missing variables in the form of &{variable_name}
-    # kwargs is a dictionary with variable names as keys and their values as values
-    # Example:
-    # !SP &{basis} PBE
-    # *xyzfile 0 1 aspirin.xyz
-    # with variables={'basis': 'def2-SVP'}
-    # will be converted to 
-    # !SP def2-SVP PBE
-    # *xyzfile 0 1 aspirin.xyz
+    """
+    Template is an input file with missing variables in the form of &{variable_name}
+    kwargs is a dictionary with variable names as keys and their values as values
+    ## Examples:
+    
+    !SP &{basis} PBE
+    *xyzfile 0 1 aspirin.xyz
+    
+    with variables={'basis': 'def2-SVP'}
+    will be converted to 
+    
+    !SP def2-SVP PBE
+    *xyzfile 0 1 aspirin.xyz
 
-    # Example with template file:
-    # tester = OrcaInputFile(OrcaInputTemplate.BASIC
-    #   , calculation='OPT'
-    #   , basis='def2-SVP'
-    #   , functional='PBE'
-    #   , xyzfile='aspirin.xyz'
-    #)
-    #Name: str
-    """Name of the Input File"""
+    ## Example with template file:
+    
+    tester = OrcaInputFile(OrcaInputTemplate.BASIC
+      , calculation='OPT'
+      , basis='def2-SVP'
+      , functional='PBE'
+      , xyzfile='aspirin.xyz'
+    )
+    """
     
     InputFileContents: str
     """The Generated Input File as a String"""
