+++ qchem/Calculation/__init__.py
@@ -1,0 +1,12 @@
+from .ClusterCalculation import ClusterCalculation
+from .OrcaCalculation import OrcaCalculation
+from .OrcaInputFile import OrcaInputFile
+from .GeoOpt import GeoOpt
+
+# Expose all Classes when importing with star (*)
+__all__ = [
+    "ClusterCalculation",
+    "OrcaCalculation",
+    "OrcaInputFile",
+    "GeoOpt"
+]
