--- qchem/Pipelines/__init__.py
@@ -1,1 +1,0 @@
-from .GeoOpt import GeoOpt
